webapp/requirements.txt
Flask==0.10.1
Flask-Scss==0.3
Jinja2==2.7.3
Markdown==2.5
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
argh==0.25.0
argparse==1.2.1
boto==2.32.1
gunicorn==19.1.1
itsdangerous==0.24
pathtools==0.1.2
pyScss==1.2.1
requests==2.4.1
six==1.8.0
watchdog==0.8.1
wsgiref==0.1.2
